[Refactor] Move JSON parser and simple path helpers to Formats and Types
