+++ be/src/exec/CMakeLists.txt
@@ -313,7 +313,6 @@ set(EXEC_NON_PIPELINE_FILES
     enforce_unique_row_locator_node.cpp
     intersect_hash_set.cpp
     intersect_node.cpp
-    json_parser.cpp
     project_node.cpp
     dict_decode_node.cpp
     repeat_node.cpp
