+++ be/src/exec/file_scanner/avro_scanner.h
@@ -24,8 +24,8 @@
 #include "compute_env/load/stream_load_pipe.h"
 #include "exec/file_scanner/file_scanner.h"
 #include "exec/file_scanner/json_scanner.h"
-#include "exprs/json_functions.h"
 #include "fs/fs.h"
+#include "types/simple_json_path.h"
 #ifdef __cplusplus
 extern "C" {
 #endif
