+++ be/test/types/CMakeLists.txt
@@ -25,6 +25,7 @@ set(TYPES_TEST_FILES
     int128_arithmetic_ops_test.cpp
     json_value_test.cpp
     logical_type_test.cpp
+    simple_json_path_test.cpp
     type_descriptor_test.cpp
     tdigest_test.cpp
     timestamp_value_test.cpp
